+++ frontend/src/contexts/AuthContext.tsx
@@ -1,5 +1,4 @@
-import { createContext, useState, useContext } from "react";
-import type { ReactNode } from "react";
+import { createContext } from "react";
 
 interface AuthContextType {
   token: string | null;
@@ -8,32 +7,5 @@ interface AuthContextType {
   logout: () => void;
 }
 
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
+export const AuthContext = createContext<AuthContextType | undefined>(undefined);
 
-export const AuthProvider = ({ children }: { children: ReactNode }) => {
-  const [token, setToken] = useState<string | null>(
-    sessionStorage.getItem("authToken")
-  );
-
-  const login = (newToken: string) => {
-    sessionStorage.setItem("authToken", newToken);
-    setToken(newToken);
-  };
-
-  const logout = () => {
-    sessionStorage.removeItem("authToken");
-    setToken(null);
-  };
-
-  return (
-    <AuthContext.Provider value={{ token, isLoggedIn: !!token, login, logout }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export const useAuth = () => {
-  const context = useContext(AuthContext);
-  if (!context) throw new Error("useAuth must be used within AuthProvider");
-  return context;
-};
